Move katsdpingest up a directory level
Now that katsdpcam2telstate, katsdpbfingest and digitiser_capture have
moved out, it was entirely in a subdirectory.

The only real changes apart from the file renames are to Jenkinsfile
(simplified slightly, although it's mostly just indentation) and
.gitignore (merged child and parent versions).
