+++ Jenkinsfile
@@ -10,27 +10,25 @@ katsdp.setDependencies([
     'ska-sa/katsdptelstate/master',
     'ska-sa/katdal'])
 
-if (!katsdp.isAborted()) {
-    catchError {
-        katsdp.stagePrepare(subdir: 'katsdpingest', python2: false, python3: true,
-                            timeout: [time: 60, unit: 'MINUTES'])
-        katsdp.stageNosetestsGpu(subdir: 'katsdpingest', cuda: true, opencl: true)
-        katsdp.stageFlake8(subdir: 'katsdpingest')
-        katsdp.stageMypy(subdir: 'katsdpingest')
-        katsdp.stageMakeDocker(subdir: 'katsdpingest', venv: true)
+catchError {
+    katsdp.stagePrepare(python2: false, python3: true,
+                        timeout: [time: 60, unit: 'MINUTES'])
+    katsdp.stageNosetestsGpu(cuda: true, opencl: true)
+    katsdp.stageFlake8()
+    katsdp.stageMypy()
+    katsdp.stageMakeDocker(venv: true)
 
-        stage('katsdpingest/autotuning') {
-            if (katsdp.notYetFailed()) {
-                katsdp.simpleNode(label: 'cuda-GeForce_GTX_TITAN_X') {
-                    deleteDir()
-                    katsdp.unpackGit()
-                    katsdp.unpackVenv()
-                    katsdp.unpackKatsdpdockerbase()
-                    katsdp.virtualenv('venv3') {
-                        dir('git/katsdpingest') {
-                            lock("katsdpingest-autotune-${env.BRANCH_NAME}") {
-                                sh './jenkins-autotune.sh titanx'
-                            }
+    stage('katsdpingest/autotuning') {
+        if (katsdp.notYetFailed()) {
+            katsdp.simpleNode(label: 'cuda-GeForce_GTX_TITAN_X') {
+                deleteDir()
+                katsdp.unpackGit()
+                katsdp.unpackVenv()
+                katsdp.unpackKatsdpdockerbase()
+                katsdp.virtualenv('venv3') {
+                    dir('git') {
+                        lock("katsdpingest-autotune-${env.BRANCH_NAME}") {
+                            sh './jenkins-autotune.sh titanx'
                         }
                     }
                 }
