--- katsdpingest/.gitignore
@@ -1,1 +1,0 @@
-.mypy_cache
