+++ katsdpingest/ingest_kernels/compress_weights.mako
