+++ katsdpingest/ingest_kernels/count_flags.mako
