+++ katsdpingest/ingest_kernels/prepare_flags.mako
